feat(hill: debug): upgrade /servers command to /debug uploads list of connected servers and quotas files

+++ hill.py
@@ -24,8 +24,9 @@ os.chdir(dname)
 load_dotenv()
 TOKEN = os.getenv('HILL_TOKEN')
 
-# Load admin's ID
+# Load admin's ID and test server ID
 admin_id = 740098404688068641
+test_server_id = 958651015064854548
 
 # Uncomment when running on replit (1/2)
 # from keep_alive import keep_alive
@@ -236,27 +237,43 @@ async def list_autocomplete(
             ][0: 25]
     return data
 
-# "servers" command
-# List all servers the bot is in
-@bot.tree.command(description="List all servers Hill is in!", guild=discord.Object(958651015064854548))
-async def servers(interaction: discord.Interaction) -> None:
+# "debug" command (test server only)
+# Upload bot and quota status to Discord for remote debug
+@bot.tree.command(description="Upload bot and quota status for debug!", guild=discord.Object(test_server_id))
+async def debug(interaction: discord.Interaction, file: str) -> None:
     await interaction.response.defer(thinking=True)
 
-    # Put list of guilds into txt file
-    guild_list_file = open('guild_list.txt', 'w')
-    for guild in bot.guilds:
-        guild_list_file.write(f'{guild.name}(id: {guild.id})\n')
-    guild_list_file.close()
+    # Send list of connected servers
+    if file == "guild_list":
+        # Put list of guilds into txt file
+        guild_list_file = open('guild_list.txt', 'w')
+        for guild in bot.guilds:
+            guild_list_file.write(f'{guild.name}(id: {guild.id})\n')
+        guild_list_file.close()
 
-    # Send the txt file
-    await interaction.edit_original_response(content=f"🎏 Bot is connected to {len(bot.guilds)} servers!")
-    await interaction.channel.send(file=discord.File("guild_list.txt"))
+        # Send the txt file
+        await interaction.edit_original_response(content=f"🎏 Bot is connected to {len(bot.guilds)} servers!")
+        await interaction.channel.send(file=discord.File("guild_list.txt"))
 
-    # Delete the txt file
-    try:
-        os.remove("guild_list.txt")
-    except:
-        pass
+        # Delete the txt file
+        try:
+            os.remove("guild_list.txt")
+        except:
+            pass
+    
+    # Send quotas file scraped by the bot
+    else:
+        await interaction.edit_original_response(content=f"🎏 Quotas file: `{file}.json`")
+        await interaction.channel.send(file=discord.File(f"{file}.json"))
+
+@debug.autocomplete('file')
+async def debug_autocomplete(
+    interaction: discord.Interaction,
+    current: str
+) -> typing.List[app_commands.Choice[str]]:
+    file_list = ['quotas', 'quotas_old', 'guild_list']
+    file_list = [app_commands.Choice(name=file, value=file) for file in file_list]
+    return file_list
 
 # Slash commands end
